+++ engine/flink/components/base/src/test/scala/pl/touk/nussknacker/engine/flink/util/transformer/ForEachTransformerSpec.scala
@@ -1,6 +1,5 @@
 package pl.touk.nussknacker.engine.flink.util.transformer
 
-import cats.data.NonEmptyList
 import com.typesafe.config.ConfigFactory
 import com.typesafe.config.ConfigValueFactory.fromAnyRef
 import org.apache.flink.streaming.api.scala._
@@ -9,27 +8,25 @@ import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.api.{ProcessListener, ProcessVersion}
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
-import pl.touk.nussknacker.engine.compiledgraph.part.PotentiallyStartPart
 import pl.touk.nussknacker.engine.deployment.DeploymentData
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.flink.util.source.EmitWatermarkAfterEachElementCollectionSource
 import pl.touk.nussknacker.engine.graph.EspProcess
 import pl.touk.nussknacker.engine.process.ExecutionConfigPreparer
-import pl.touk.nussknacker.engine.process.compiler.{FlinkProcessCompiler, UsedNodes}
+import pl.touk.nussknacker.engine.process.compiler.FlinkProcessCompiler
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
 import pl.touk.nussknacker.engine.spel.Implicits._
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.testmode._
 
 import java.time.Duration
-import java.util.UUID
 
 class ForEachTransformerSpec extends FunSuite with FlinkSpec with Matchers with Inside {
 
   private val sinkId = "end"
   private val resultVariableName = "resultVar"
   private val forEachOutputVariableName = "forEachVar"
-  private val forEachNodeId = "for-each"
+  private val forEachNodeResultId = "for-each-result"
 
   test("should produce results for each element in list") {
     val collectingListener = initializeListener
@@ -47,12 +44,8 @@ class ForEachTransformerSpec extends FunSuite with FlinkSpec with Matchers with 
 
     val testProcess = aProcessWithForEachNode(elements = "{'one', 'other'}", resultExpression = s"#$forEachOutputVariableName + '_1'")
 
-    val compiledSources: NonEmptyList[PotentiallyStartPart] = getCompiledSources(model, testProcess)
-
-    val forEachOutputVariable = compiledSources.head.nextParts.find(_.id == forEachNodeId)
-      .get.validationContext.variables(forEachOutputVariableName)
-
-    forEachOutputVariable shouldBe Typed[String]
+    val forEachResultValidationContext = model.validator.validate(testProcess).typing(forEachNodeResultId)
+    forEachResultValidationContext.inputValidationContext.get(forEachOutputVariableName) shouldBe Some(Typed[String])
   }
 
   test("should not produce any results when elements list is empty") {
@@ -76,8 +69,8 @@ class ForEachTransformerSpec extends FunSuite with FlinkSpec with Matchers with 
       .parallelism(1)
       .stateOnDisk(true)
       .source("start", "start")
-      .customNode(forEachNodeId, forEachOutputVariableName, "for-each", "Elements" -> elements)
-      .buildSimpleVariable("for-each-result", "resultVar", resultExpression)
+      .customNode("for-each", forEachOutputVariableName, "for-each", "Elements" -> elements)
+      .buildSimpleVariable(forEachNodeResultId, "resultVar", resultExpression)
       .emptySink(sinkId, "dead-end")
 
   private def collectTestResults[T](model: LocalModelData, testProcess: EspProcess, collectingListener: ResultsCollectingListener): TestProcess.TestResults[Any] = {
@@ -87,14 +80,6 @@ class ForEachTransformerSpec extends FunSuite with FlinkSpec with Matchers with 
 
   private def extractResultValues(results: TestProcess.TestResults[Any]): List[String] = results.nodeResults(sinkId)
     .map(_.variableTyped[String](resultVariableName).get)
-
-  private def getCompiledSources(model: LocalModelData, testProcess: EspProcess): NonEmptyList[PotentiallyStartPart] = {
-    val compiler = new FlinkProcessCompiler(model)
-    val compiledSources = compiler.compileProcess(testProcess, ProcessVersion.empty, DeploymentData.empty, new TestServiceInvocationCollector(TestRunId(UUID.randomUUID().toString)))(UsedNodes.empty, getClass.getClassLoader)
-      .compileProcess()
-      .sources
-    compiledSources
-  }
 
   private def runProcess(model: LocalModelData, testProcess: EspProcess): Unit = {
     val stoppableEnv = flinkMiniCluster.createExecutionEnvironment()