+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/functional/MetricsSpec.scala
@@ -1,14 +1,19 @@
 package pl.touk.nussknacker.engine.process.functional
 
+import cats.data.NonEmptyList
+
 import java.util.Date
 import org.apache.flink.configuration.Configuration
 import org.scalatest.{BeforeAndAfterEach, FunSuite, Matchers}
 import org.scalatest.LoneElement._
+import pl.touk.nussknacker.engine.api.{MetaData, StreamMetaData}
 import pl.touk.nussknacker.engine.build.{GraphBuilder, ScenarioBuilder}
-import pl.touk.nussknacker.engine.graph.node.Case
+import pl.touk.nussknacker.engine.graph.EspProcess
+import pl.touk.nussknacker.engine.graph.node.{Case, DeadEndingData, EndingNodeData}
 import pl.touk.nussknacker.engine.process.helpers.ProcessTestHelpers
 import pl.touk.nussknacker.engine.process.helpers.SampleNodes.{MockService, SimpleRecord, SinkForStrings}
 import pl.touk.nussknacker.engine.spel
+import pl.touk.nussknacker.engine.split.NodesCollector
 import pl.touk.nussknacker.test.VeryPatientScalaFutures
 
 class MetricsSpec extends FunSuite with Matchers with VeryPatientScalaFutures with ProcessTestHelpers with BeforeAndAfterEach {
@@ -88,14 +93,12 @@ class MetricsSpec extends FunSuite with Matchers with VeryPatientScalaFutures wi
 
     processInvoker.invokeWithSampleData(process, data)
 
-    //eventually {
-      counter("nodeId.source1.nodeCount") shouldBe 2L
-      counter("nodeId.filter1.nodeCount") shouldBe 2L
-      counter("nodeId.split1.nodeCount") shouldBe 1L
-      counter("nodeId.proc2.nodeCount") shouldBe 1L
-      counter("nodeId.out.nodeCount") shouldBe 1L
-      counter("nodeId.out2.nodeCount") shouldBe 1L
-    //}
+    counter("nodeId.source1.nodeCount") shouldBe 2L
+    counter("nodeId.filter1.nodeCount") shouldBe 2L
+    counter("nodeId.split1.nodeCount") shouldBe 1L
+    counter("nodeId.proc2.nodeCount") shouldBe 1L
+    counter("nodeId.out.nodeCount") shouldBe 1L
+    counter("nodeId.out2.nodeCount") shouldBe 1L
   }
 
   test("measure ends") {
@@ -115,16 +118,14 @@ class MetricsSpec extends FunSuite with Matchers with VeryPatientScalaFutures wi
 
     processInvoker.invokeWithSampleData(process, data)
 
-    //eventually {
-      counter("end.nodeId.sink.count") shouldBe 1L
-      gauge("end.nodeId.sink.instantRate") should be >= 0.0d
+    counter("end.nodeId.sink.count") shouldBe 1L
+    gauge("end.nodeId.sink.instantRate") should be >= 0.0d
 
-      counter("end.nodeId.processor.count") shouldBe 1L
-      gauge("end.nodeId.processor.instantRate") should be >= 0.0d
+    counter("end.nodeId.processor.count") shouldBe 1L
+    gauge("end.nodeId.processor.instantRate") should be >= 0.0d
 
-      counter("end.nodeId.custom node.count") shouldBe 1L
-      gauge("end.nodeId.custom node.instantRate") should be >= 0.0d
-    //}
+    counter("end.nodeId.custom node.count") shouldBe 1L
+    gauge("end.nodeId.custom node.instantRate") should be >= 0.0d
   }
 
   test("measure dead ends") {
@@ -144,12 +145,10 @@ class MetricsSpec extends FunSuite with Matchers with VeryPatientScalaFutures wi
 
     processInvoker.invokeWithSampleData(process, data)
 
-    //eventually {
-      counter("dead_end.nodeId.filter1.count") shouldBe 1L
-      gauge("dead_end.nodeId.filter1.instantRate") should be >= 0.0d
-      counter("dead_end.nodeId.switch2.count") shouldBe 2L
-      gauge("dead_end.nodeId.switch2.instantRate") should be >= 0.0d
-    //}
+    counter("dead_end.nodeId.filter1.count") shouldBe 1L
+    gauge("dead_end.nodeId.filter1.instantRate") should be >= 0.0d
+    counter("dead_end.nodeId.switch2.count") shouldBe 2L
+    gauge("dead_end.nodeId.switch2.instantRate") should be >= 0.0d
   }
 
   test("open measuring service"){
@@ -163,16 +162,50 @@ class MetricsSpec extends FunSuite with Matchers with VeryPatientScalaFutures wi
     )
 
     processInvoker.invokeWithSampleData(process, data)
-    eventually {
-      SinkForStrings.data shouldBe List("initialized!")
+    SinkForStrings.data shouldBe List("initialized!")
+  }
+
+  test("initializes counts, ends, dead ends") {
+
+    val scenario = EspProcess(MetaData("proc1", StreamMetaData()), NonEmptyList.of(
+      GraphBuilder.source("id", "input")
+        .split("split",
+          GraphBuilder.filter("left", "false").branchEnd("end1", "join1"),
+          GraphBuilder.filter("right", "false").branchEnd("end2", "join1")
+        ),
+      GraphBuilder.join("join1", "joinBranchExpression", Some("any"),
+        List(
+          "end1" -> List("value" -> "''"),
+          "end2" -> List("value" -> "''")
+        ))
+        .customNodeNoOutput("custom", "customFilter", "input" -> "''", "stringVal" -> "''")
+        .processor("proc1", "lifecycleService")
+        .switch("switch1", "false", "any2",
+          GraphBuilder.emptySink("outE1", "sinkForStrings", "value" -> "''"),
+          Case("true", GraphBuilder.processorEnd("procE1", "lifecycleService")),
+          Case("false", GraphBuilder.endingCustomNode("customE1", None,"optionalEndingCustom", "param" -> "''"))
+        )
+    ))
+    val allNodes = NodesCollector.collectNodesInScenario(scenario).map(_.data)
+
+    processInvoker.invokeWithSampleData(scenario, Nil)
+
+    allNodes.foreach { node =>
+      counter(s"nodeId.${node.id}.nodeCount") shouldBe 0L
+    }
+    allNodes.filter(_.isInstanceOf[EndingNodeData]).foreach { node =>
+      counter(s"end.nodeId.${node.id}.count") shouldBe 0L
+    }
+    allNodes.filter(_.isInstanceOf[DeadEndingData]).foreach { node =>
+      counter(s"dead_end.nodeId.${node.id}.count") shouldBe 0L
     }
   }
 
-  private def counter(name: String): Long = {
+  private def counter(name: String): Long = withClue(s"counter $name") {
     TestReporter.get(this.getClass.getName).testCounters(name).loneElement.getCount
   }
 
-  private def gauge(name: String): Double = {
+  private def gauge(name: String): Double = withClue(s"gauge $name"){
     TestReporter.get(this.getClass.getName).testGauges(name).loneElement.getValue.asInstanceOf[Double]
   }